*: implement a v1beta2 API, support WI
Introducing a new v1beta2 API for a number of reasons:

 - the v1beta1 API made a number of choices that lead to less secure and
   clear systems, like:
  - implicit behavior with nominally required but unset fields being
    defaulted from hard-coded values in the controller or from the VMSS
    or from the controller context. We want security-sensitive
    functionality to require users to be explicit and opt into
    specifically the behavior they need. Might it be a bit more work?
    Sure, but it's always worth it.
  - allowing things like ACR token scope to be omitted and providing the
    user with a raw Entra token is bad practice
 - the v1beta1 API did not provide for future-proofing the API surface,
   amalgamating all of the fields in one top-level container; adding all
   of the  workload identity fields here would have made validation
   logic challenging to write
 - the v1beta1 API did not provide strict validation for inputs

All of these required breaking changes to the API to fix; it is much
easier to release a new version of the API and ask users to fail forward
than it would be to do some complex versioning dance.

The new API, along with all of these security and quality-of-life
improvements, supports the use of workload identity for generating these
image pull secrets.

Signed-off-by: Steve Kuznetsov <stekuznetsov@microsoft.com>